==> analysis/01_first-successful-import-and-prediction.Rmd <==
---
title: "Getting the Hong Kong Data into R and making a random forest"
output: github_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

libraries 

```{r}
library(randomForest)
library(raster)
library(rgdal)
library(here)
library(tidyverse)
library(mapproj)
#training <- readOGR('training_data.shp', layer='training_data')
train <- brick(here("data/hong_kong/lcz/", "hong_kong_lcz_gt.tif"))
#le7 <- brick('LE70220491999322EDC01_stack.gtif')
le8 <- brick(here("data/hong_kong/landsat_8", "LC81220442015291LGN00/LC81220442015291LGN00_B1.tif"))
```


```{r}
all_data <- stack(le8, train)
```



```{r}
summary(all_data)
```

Convert all into dataframes

```{r}
all_data_df <- as.data.frame(all_data, xy=TRUE)
```


```{r}
le8_df <- as.data.frame(le8, xy=TRUE)
train_df <- as.data.frame(train, xy=TRUE)
```

Compare three different dfs to see how `stack()` put things together

```{r}
summary(all_data_df)
summary(le8_df)
summary(train_df)
```





```{r}
train_df %>%
  filter(layer !=0) %>%
  ggplot()+
  geom_histogram(aes(x=layer))
```

```{r}
train_df %>%
  filter(layer !=0) %>%
  mutate(lcz = factor(layer)) %>%
  summary()
```




```{r}
ggplot() +
    geom_raster(data = le8_df , aes(x = x, y = y, fill = layer)) +
    scale_fill_viridis_c() +
    coord_quickmap() +
  theme(legend.position = "none")
```


## 1/14/21

#### Is GT what I think it is? 

I think that I've figured out how to match up my pixels with theirs, but some of the numbers seem off so I'm going to compare them here: 

```{r}
table_3 <- data.frame(
  lcz =  c(1:17),
  training = c(318, 112, 195, 383, 76, 64, 0, 86, 0, 109, 832, 207, 379, 332, 0, 0, 1282),
  test = c(313, 67, 131, 290, 50, 56, 0, 51, 0, 110, 784, 200, 312, 236, 0, 0, 1054)
)
table_3 <- table_3 %>%
  mutate(total_pixels = training+test)
```

```{r}
summary(factor(train_df$layer))
table_3$total_pixels
```

So they're all the same in the table from the paper as in my imported data except for 12, 13, 14, and 17, each of which has more in my imported data than in the table. Did they clip the data more closely to Hong Kong's limits?

12, 13, 14, 17 are scattered trees, bush/scrub, low plants, and water, respectively. So it would make sense if it's just a difference in outer limit. Not sure how to figure out exactly how they clipped the data, but I'm going to put that off for rn. 

```{r}
rm(all_data, all_data_df, le8, le8_df, train, train_df)
```


#### Bands

They use bands 1-7, 10, and 11. 

I'm going to just pick a day and get those bands in. 

```{r}
lcz_train <- brick(here("data/hong_kong/lcz/", "hong_kong_lcz_gt.tif"))
b1 <- brick(here("data/hong_kong/landsat_8", "LC81220442015291LGN00/LC81220442015291LGN00_B1.tif"))
b2 <- brick(here("data/hong_kong/landsat_8", "LC81220442015291LGN00/LC81220442015291LGN00_B2.tif"))
b3 <- brick(here("data/hong_kong/landsat_8", "LC81220442015291LGN00/LC81220442015291LGN00_B3.tif"))
b4 <- brick(here("data/hong_kong/landsat_8", "LC81220442015291LGN00/LC81220442015291LGN00_B4.tif"))
b5 <- brick(here("data/hong_kong/landsat_8", "LC81220442015291LGN00/LC81220442015291LGN00_B5.tif"))
b6 <- brick(here("data/hong_kong/landsat_8", "LC81220442015291LGN00/LC81220442015291LGN00_B6.tif"))
b7 <- brick(here("data/hong_kong/landsat_8", "LC81220442015291LGN00/LC81220442015291LGN00_B7.tif"))
b10 <- brick(here("data/hong_kong/landsat_8", "LC81220442015291LGN00/LC81220442015291LGN00_B10.tif"))
b11 <- brick(here("data/hong_kong/landsat_8", "LC81220442015291LGN00/LC81220442015291LGN00_B11.tif"))
```

Put them all together

```{r}
lcz_and_bands <- stack(lcz_train, b1, b2, b3, b4, b5, b6, b7, b10, b11)
```

change to df and fix column names

```{r}
lcz_and_bands_df <- as.data.frame(lcz_and_bands)
names(lcz_and_bands_df) <- c("lcz", "b1", "b2", "b3", "b4", "b5", "b6", "b7", "b10", "b11")
```

Does it look right? 

```{r}
summary(lcz_and_bands_df)
```

Yes. Let's trim it down now and make lcz a factor

```{r}
rf_training_df <- lcz_and_bands_df %>%
  filter(lcz != 0) %>%
  mutate(lcz = factor(lcz))

summary(rf_training_df)
```

What are bands 10 and 11 and why do they have so many NAs?
They're the "30m resolution thermal bands interpolated from 100 m resolution data collected from Thermal Infrared Sensor (TIRS)"
I'm assuming I need to cut them out? Going to try. Maybe it'll make the pixel numbers match their table 

```{r}
rf_training_df_nona <- rf_training_df %>%
  drop_na()

summary(rf_training_df_nona)
```

Let's compare with table 3

```{r}
summary(rf_training_df_nona$lcz)
table_3$total_pixels
```

Well now only 14 (low plants) and 17 (water) have too many, which is something. 
When I look at it in QGIS there's one 17 classed polygon that goes outside the limits, but I don't see any class 14. 
This is too much of a sidetrack.

#### Fit a random forest (?!)

```{r}
rf <- randomForest(lcz ~ b1 + b2 + b3 + b4 + b5 + b6 + b7 + b10 +b11, data=rf_training_df_nona, importance=TRUE)
print(rf)
```

```{r}
importance(rf)
```


```{r}
varImpPlot(rf)
```

Bring back just the landsat data

```{r}
only_bands <- stack(b1, b2, b3, b4, b5, b6, b7, b10, b11)
names(only_bands) <- c('b1', 'b2', 'b3', 'b4', 'b5', 'b6', 'b7', 'b10', 'b11')
```

Do the actual prediction: 

```{r}
bands_lcz_predicted <- predict(only_bands, model=rf, na.rm=T)
```

**I got a bunch of errors when I tried to make the prediction with only_bands in a data frame (rather than a RasterStack) and that doesn't make sense to me since the random forest was created with a data frame.**

Make a plot: 

```{r}
plot(bands_lcz_predicted)
```

Save this file to look at in QGIS

```{r}
writeRaster(bands_lcz_predicted, here("results", "first_successful_prediction_01142021.tif"), overwrite=T)
```
