==> analysis/03_initial-data-cleaning.Rmd <==
---
title: "Initial Data Cleaning"
output: github_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
library(tidyverse)
library(raster)
library(rgdal)
library(here)
library(magrittr)
source(here("R", "get_rasterstack.R"))
```

Strategy is to load all scenes in as rasterstacks and the lcz and polygon information as a dataframe.

The scenes are then stacked and converted to a dataframe.

LCZ and scene data is bound by column and saved in `results` as `hk_df.rds` file

```{r load-data}
lcz <- readRDS(here("results", "lcz_and_poly_ids.RDS"))
scenes <- get_rasterstack()
```

```{r}
scenes_df <- as.data.frame(scenes)
```

```{r put-it-all-together}
lcz_and_scenes_df <- bind_cols(lcz, scenes_df)
```

```{r fix-data-types}
names(lcz_and_scenes_df)[1] <- "lcz"
```


```{r save}
saveRDS(lcz_and_scenes_df, here("results", "hk_df.rds"))
```
