==> analysis/05_different-classification-schemes.Rmd <==
---
title: "DIfferent Classification Schemes"
author: "Ericka B. Smith"
date: "2/1/2021"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
library(randomForest)
library(tidyverse)
library(here)
library(raster)
library(rgdal)
library(mapproj)
source(here("R", "get_oa_metrics.R"))
```

## Load Data

```{r}
train <- readRDS(here("results", "train.rds"))
test <- readRDS(here("results", "test.rds"))
```

# Start with default mtry=6

## Create & Evaluate Model

```{r}
rf <- randomForest(lcz ~ ., data=train, importance=TRUE)
print(rf)
```

```{r}
# importance(rf)
```

```{r, fig.height=12, fig.width=12}
varImpPlot(rf)
```

## Predict

```{r}
only_bands <- test[,2:37]
```

```{r}
bands_lcz_predicted <- predict(rf, newdata=only_bands, na.rm=T)
```

```{r}
test1 <- test
test1$lcz_predicted <- bands_lcz_predicted
```


## Check Predictions

```{r}
test1 %>%
  group_by(lcz, lcz_predicted) %>%
  summarise(n=n()) %>%
  ungroup() %>%
  filter(lcz != lcz_predicted) %>%
  ggplot()+
  geom_tile(aes(x=lcz, y=lcz_predicted, fill=n))+
  scale_fill_viridis_c(option="plasma")+
  labs(title= "Incorrect LCZ Predictions",caption = "Note that class 17 (Water) is not included on the x-axis.\nThis is because all predictions for this group were correct.")+
  xlab("Actual LCZ Class")+
  ylab("Predicted LCZ Class")+
  theme_classic()+
  theme(axis.ticks = element_blank())
```


<!-- ## Predict with all data and save as raster (just because) -->

<!-- ```{r} -->
<!-- all_hk_bands <- stackOpen(here("results", "bands_rasterstack.stk")) -->
<!-- names(all_hk_bands) <- names(only_bands) -->
<!-- ``` -->

<!-- ```{r} -->
<!-- all_hk_bands_lcz_predicted <- predict(all_hk_bands, model=rf, na.rm=T) -->
<!-- ``` -->

<!-- ```{r} -->
<!-- plot(all_hk_bands_lcz_predicted) -->
<!-- ``` -->

<!-- ```{r} -->
<!-- writeRaster(all_hk_bands_lcz_predicted, here("results", "hk_pred1_02012021.tif"), overwrite=T) -->
<!-- ``` -->


# Again with mtry=8

```{r}
rf2 <- randomForest(lcz ~ ., data=train, importance=TRUE, mtry=8)
print(rf2)
```

```{r}
# importance(rf2)
```

```{r, fig.height=12, fig.width=12}
varImpPlot(rf2)
```

## Predict

```{r}
only_bands2 <- test[,2:37]
```

```{r}
bands_lcz_predicted2 <- predict(rf2, newdata=test, na.rm=T)
```

```{r}
test2 <- test
test2$lcz_predicted <- bands_lcz_predicted2
```


## Check Predictions


```{r}
test2 %>%
  group_by(lcz, lcz_predicted) %>%
  summarise(n=n()) %>%
  ungroup() %>%
  filter(lcz != lcz_predicted) %>%
  ggplot()+
  geom_tile(aes(x=lcz, y=lcz_predicted, fill=n))+
  scale_fill_viridis_c(option="plasma")+
  labs(title= "Incorrect LCZ Predictions for rf2")+
  xlab("Actual LCZ Class")+
  ylab("Predicted LCZ Class")+
  theme_classic()+
  theme(axis.ticks = element_blank())
```


# Again with mtry=4


```{r}
rf3 <- randomForest(lcz ~ ., data=train, importance=TRUE, mtry=4)
print(rf3)
```


```{r}
# importance(rf2)
```


```{r, fig.height=12, fig.width=12}
varImpPlot(rf3)
```

## Predict

```{r}
only_bands3 <- test[,2:37]
```


```{r}
bands_lcz_predicted3 <- predict(rf3, newdata=test, na.rm=T)
```


```{r}
test3 <- test
test3$lcz_predicted <- bands_lcz_predicted3
```


## Check Predictions


```{r}
test3 %>%
  group_by(lcz, lcz_predicted) %>%
  summarise(n=n()) %>%
  ungroup() %>%
  filter(lcz != lcz_predicted) %>%
  ggplot()+
  geom_tile(aes(x=lcz, y=lcz_predicted, fill=n))+
  scale_fill_viridis_c(option="plasma")+
  labs(title= "Incorrect LCZ Predictions for rf3")+
  xlab("Actual LCZ Class")+
  ylab("Predicted LCZ Class")+
  theme_classic()+
  theme(axis.ticks = element_blank())
```


# Compare all 3 with different mtry

```{r}
get_oa_metrics(test1)
get_oa_metrics(test2)
get_oa_metrics(test3)
```


# change ntrees to 1000, mtry back to 6


```{r}
rf4 <- randomForest(lcz ~ ., data=train, importance=TRUE, ntree=1000)
print(rf4)
```


```{r}
# importance(rf2)
```


```{r, fig.height=12, fig.width=12}
varImpPlot(rf4)
```

## Predict

```{r}
only_bands4 <- test[,2:37]
```


```{r}
bands_lcz_predicted4 <- predict(rf4, newdata=test, na.rm=T)
```


```{r}
test4 <- test
test4$lcz_predicted <- bands_lcz_predicted4
```


## Check Predictions


```{r}
test4 %>%
  group_by(lcz, lcz_predicted) %>%
  summarise(n=n()) %>%
  ungroup() %>%
  filter(lcz != lcz_predicted) %>%
  ggplot()+
  geom_tile(aes(x=lcz, y=lcz_predicted, fill=n))+
  scale_fill_viridis_c(option="plasma")+
  labs(title= "Incorrect LCZ Predictions for rf4")+
  xlab("Actual LCZ Class")+
  ylab("Predicted LCZ Class")+
  theme_classic()+
  theme(axis.ticks = element_blank())
```


# Now with ntree=250, mtry still 6


```{r}
rf5 <- randomForest(lcz ~ ., data=train, importance=TRUE,ntree=250)
print(rf5)
```


```{r}
# importance(rf2)
```


```{r, fig.height=12, fig.width=12}
varImpPlot(rf5)
```

## Predict

```{r}
only_bands5 <- test[,2:37]
```


```{r}
bands_lcz_predicted5 <- predict(rf5, newdata=test, na.rm=T)
```


```{r}
test5 <- test
test5$lcz_predicted <- bands_lcz_predicted5
```


## Check Predictions


```{r}
test5 %>%
  group_by(lcz, lcz_predicted) %>%
  summarise(n=n()) %>%
  ungroup() %>%
  filter(lcz != lcz_predicted) %>%
  ggplot()+
  geom_tile(aes(x=lcz, y=lcz_predicted, fill=n))+
  scale_fill_viridis_c(option="plasma")+
  labs(title= "Incorrect LCZ Predictions for rf5")+
  xlab("Actual LCZ Class")+
  ylab("Predicted LCZ Class")+
  theme_classic()+
  theme(axis.ticks = element_blank())
```


# Compare all 3 with different ntree

```{r}
get_oa_metrics(test1)
get_oa_metrics(test4)
get_oa_metrics(test5)
```
