==> analysis/08_tuning-parameter-plots.Rmd <==
---
title: "08_tuning-parameter-plots"
author: "Ericka Smith"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
library(tidyverse)
library(here)
library(forcats)
library(reshape2)
library(viridis)
library(ggpubr)
library(cowplot)
theme_set(theme_pubr())
```

Load data

```{r, warning=FALSE}
current_data_name <- "ntree_5_to_500_data.rds"
param_dat <-
  read_csv(here("results/run_output_tables", current_data_name))
param_dat <- param_dat[, 2:18]
```

Name output plots

```{r}
current_line_plot_name <- "OA_line_plot.png"
plots_save_location <- "doc/presentation/images"
f1_facet_plot_name <- "f1_facet.png"
```

```{r fig.height=9, fig.width=12}
param_dat %>%
  dplyr::select(ntree, oa, oa_urb, oa_nat) %>%
  melt(
    id.vars = "ntree",
    measure.vars = c("oa", "oa_urb", "oa_nat"),
    value.name = "Accuracy",
    variable.name = "Metric"
  ) %>%
  ggplot(aes(x = ntree, y = Accuracy, color = Metric)) +
  geom_point() +
  geom_smooth(se = F) +
  labs(title = "OA Metrics increase as number of trees increase, leveling off \naround 125 trees. Natural classes have higher OA values.\n", #"Overall Accuracy Metrics by Numbers of Trees",
  x ="\nNumber of Trees",
  y = "Accuracy\n") +
  scale_color_manual(
    breaks = c("oa", "oa_urb", "oa_nat"),
    labels = c("Overall", "Urban", "Natural"),
    values = c("#FDE725FF", "#1F968BFF", "#440154FF")
  ) +
  theme(legend.position = "right",
        legend.title = element_text(size=20),
        legend.text = element_text(size=16),
        axis.title = element_text(size=16),
        title = element_text(size = 20))

ggsave(here(plots_save_location, current_line_plot_name))
```

```{r}
urban_classes <- c("1", "2", "3", "4", "5", "6", "8", "10")
f1_list <- names(param_dat)[c(1, 5:ncol(param_dat))]
# make plot
dat <- param_dat %>%
  dplyr::select(all_of(f1_list)) %>%
  rename_at(vars(contains("f1_")), funs(str_replace(., "f1_", ""))) %>%
  melt(id.vars = "ntree",
       value.name = "F1",
       variable.name = "Class") %>%
  mutate(Type = ifelse(Class %in% urban_classes, "Urban", "Natural"))
```

```{r}
# ggplot(dat) +
#   geom_raster(aes(x = ntree, y = Class, fill = F1)) +
#   scale_fill_viridis(option = "magma") +
#   labs(title = "F-1 Score by Class for 25 to 2500 Trees",
#        x = "\nNumber of Trees",
#        y = "Class\n") +
#   theme(legend.position = "right")
# 
# ggsave(here("results/plots", current_heatmap_name))
```


```{r fig.height=9, fig.width=12}
make_plot <- function(the_dat, which_type){
  the_dat %>%
  ggplot(aes(x = ntree, y = F1, color = Type)) +
  geom_smooth(se = F, size=0.75) +
  geom_point(size = 0.75)+
labs(title ="",
     x ="") +
  facet_grid(Class ~ ., scales = "free") +
  scale_color_manual(values = c("Urban" = "#1F968BFF", "Natural" = "#440154FF")) +
  theme(
    axis.text = element_text(size = 12),
    strip.text.y = element_text(angle = 0),
    legend.position = "none"
  )
}
urb_dat <- dat %>%
  filter(Type == "Urban")
urb <- make_plot(urb_dat) + theme(legend.position="none",
                                  axis.title.y = element_text(size=16, face="bold")) + labs(y = "F-1 Score\n")
nat_dat <- dat %>%
  filter(Type == "Natural")
nat <- make_plot(nat_dat) + theme(legend.position = "none")+ labs(y = "")

all <- make_plot(dat)

# extract the legend from one of the plots
legend <- get_legend(# create some space to the left of the legend
  all + theme(
    legend.position = "right",
    legend.box.margin = ggplot2::margin(0, 0, 0, 12),
    legend.title = element_text(size=20),
    legend.text = element_text(size=16))
  )


facet_pair <-
  plot_grid(urb, nat, legend,
    nrow = 1, rel_widths = c(1,0.9, 0.3)
  )
annotate_figure(facet_pair,
                top = text_grob(
                  "F-1 Score by Class increases as number of trees increases, leveling off around 100 trees.\n",
                  #"\nF-1 Score by Class for 5 to 500 Trees",
                  #face = "bold",
                  size = 20
                ),
                bottom = text_grob("Number of Trees", 
                                   face="bold",
                                   size = 16))#,
                #left = text_grob("F-1 Score", rot=90))

ggsave(here(plots_save_location, f1_facet_plot_name))
```

Try putting them together

```{r}
# param_higher_ntree_dat <-
#   read_csv(here("results/run_output_tables", "ntree_25_to_2500_data.rds"))
# param_higher_ntree_dat <- param_higher_ntree_dat[, 2:18]
# 
# dat_25_2500 <- param_higher_ntree_dat %>%
#   dplyr::select(all_of(f1_list)) %>%
#   rename_at(vars(contains("f1_")), funs(str_replace(., "f1_", ""))) %>%
#   melt(id.vars = "ntree",
#        value.name = "F1",
#        variable.name = "Class") %>%
#   mutate(Type = ifelse(Class %in% urban_classes, "Urban", "Natural"))
```


```{r fig.height=7, fig.width=5}
# p2 <- ggplot(dat_25_2500, aes(x = ntree, y = F1, color = Type)) +
#   geom_smooth(se = F, size=0.5) +
#   geom_point(size = 0.25) +
#   labs(title ="F-1 Score by Class for 25 to 2500 Trees",
#        x =  "\nNumber of Trees",
#        y = "F-1 Score") +
#   facet_grid(Class ~ ., scales = "free") +
#   scale_color_manual(values = c("Urban" = "#1F968BFF", "Natural" = "#440154FF")) +
#   theme(
#     axis.text = element_text(size = 8),
#     strip.text.y = element_text(angle = 0),
#     legend.position = "right"
#   )
# ggsave(here("results/plots", "ntree_25_to_2500_facet_plot.pdf"))

# facet_pair <- ggarrange(p1, p2, common.legend = F, legend = "bottom")
# annotate_figure(facet_pair, top = text_grob("\nF-1 Score by Class for Two Scales of Numbers of Trees", face="bold", size = 14))
# ggsave(here("results/plots", "ntree_5_to_2500_combined_facet_plot.pdf"))
```
