==> analysis/09_winning-random-forest.Rmd <==
---
title: "08_winning-random-forest"
author: "Ericka B. Smith"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
library(randomForest)
library(tidyverse)
library(here)
library(raster)
library(rgdal)
library(forcats)
library(mapproj)
library(broom)
library(ggpubr)
library(viridis)
library(cowplot)
library(magrittr)
```

```{r}
theme_set(theme_pubclean())
```

Load Data and Functions

```{r}
source(here("R", "get_oa_metrics.R"))
source(here("R", "get_f1_score.R"))
source(here("R", "predict_lcz.R"))
source(here("R", "try_parameter.R"))
train <- readRDS(here("results", "train.rds"))
test <- readRDS(here("results", "test.rds"))
scene_dat <- read_csv(here("data", "scene_info.csv"))
band_info <- read_csv(here("data", "landsat_bands.txt")) %>%
  mutate(Band = factor(Band))
```

Load Model


```{r}
rf <- readRDS(here("results", "winning_rf.rds"))
```

```{r}
#rf <- randomForest(lcz ~ ., data=train, ntree=125, importance=TRUE)
```


Predict on Test Data

```{r}
predicted <- predict_lcz(test, rf)
```

Get metrics

```{r}
oa <- get_oa_metrics(predicted)
f1 <- get_f1_score(predicted, lcz, lcz_predicted)
```


# Make Plots

```{r fig.height=9, fig.width=12}
imp_plot <- varImpPlot(rf)
class(imp_plot)

get_date <- function(scene_num){
  date <- scene_dat$Date[which(scene_dat$Scene==scene_num)]
}

band_info$Band %<>% fct_inorder()

rf$importance %>%
  as.data.frame() %>%
  rownames_to_column(var = "Predictor Variable") %>%
  separate(
    col = `Predictor Variable`,
    into = c("Band", "Scene"),
    sep = -1,
    remove = F
  ) %>%
  mutate(Band = factor(parse_number(Band)),
         Scene = as.double(Scene)) %>%
  left_join(scene_dat, by="Scene") %>%
  left_join(band_info, by="Band") %>%
  mutate(Scenes = paste("Scene", Scene, ":\n", Date),
         Band = paste("Band", Band, ":\n", Description),
         Band = fct_inorder(Band)) %>%
  ggplot() +
  geom_bar(aes(x = MeanDecreaseGini, y = Band, fill = Scenes),
           stat = 'identity',
           position = "dodge",
           color="black") +
  facet_grid(Band~., scales="free")+
  scale_fill_viridis(discrete = T) +
  theme(legend.position = "right",
        legend.title = element_blank(),
        strip.text=element_blank(),
        legend.text = element_text(size=16),
        axis.title= element_text(size = 16),
        title = element_text(size = 20),
        axis.text.y = element_text(size = 16)) +
  labs(title = "Importance Measures don't give a clear answer about \nwhich predictor variables are most useful.\n",
       x = "\nMean Decrease in Gini Impurity",
       y="")

ggsave(here("doc/presentation/images", "imp_ntree_125.png"))
```


```{r fig.height=9, fig.width=12}
urban <-
  c("F1-10",
    "F1-8",
    "F1-6",
    "F1-5",
    "F1-4",
    "F1-3",
    "F1-2",
    "F1-1",
    "OA-URB")
natural <- c("F1-17", "F1-14", "F1-13", "F1-12", "F1-11", "OA-NAT")

prep_and_plot <- function(dat, metric_name) {
  dat %>%
    rename_at(vars(everything()), ~ toupper(gsub("_", "-", .x))) %>%
    pivot_longer(everything(), names_to = "Metric") %>%
    mutate(
      Type = case_when(
        Metric %in% urban ~ "Urban",
        Metric %in% natural ~ "Natural",
        TRUE ~ "Overall"
      ),
      Metric = fct_reorder(Metric, value)
    ) %>%
    ggplot() +
    geom_bar(aes(x = value, y = Metric, fill = Type),
             color = "black",
             stat = 'identity') +
    scale_fill_viridis(discrete = T, option = "cividis") +
    theme_pubr()+
    # labs(
    #   title = "",
    #   subtitle = "",
    #   x = "",
    #   y = metric_name
    # ) +
    coord_cartesian(xlim = c(0, 1)) +
    theme(plot.margin = unit(c(0.1, 0.1, 0.1, 0.1), "cm"),
          axis.text = element_text(size = 16),
          axis.title = element_blank(),
          panel.grid.major.x = element_line(color="black", linetype="dotted"))
  
}

oa_plot <- prep_and_plot(oa, "")
f1_plot <- prep_and_plot(f1, "") + facet_grid(Type ~ ., scales = "free")

# extract the legend from one of the plots
legend <- get_legend(
  oa_plot + theme(
    legend.position = "right",
    legend.box.margin = ggplot2::margin(0, 0, 0, 0),
    legend.text = element_text(size=16),
    legend.title = element_text(size = 20)
  ))
oa_just_plot <- oa_plot + theme(legend.position = "none")
f1_just_plot <-
  f1_plot + theme(legend.position = "none", strip.text = element_blank())

left_side <- plot_grid(legend, oa_just_plot, rel_heights = c(0.4, 1), nrow=2)
left_side

facet_pair <-
  plot_grid(
    #left_side,
    f1_just_plot,
    left_side,
    ncol = 2,
    #rel_widths = c(0.75, 1)
    rel_widths = c(1, 0.75)
    
  )
annotate_figure(facet_pair,
                top = text_grob(
                  "Validation Metrics are much lower for test data than for out-of-bag data. \nHigh OA values may mask low F1 scores.\n",
                  #face = "bold",
                  size = 20
                ))
ggsave(here(
  "doc/presentation/images",
  "valid_mets.png"
))
```


Save Random Forest

```{r}
#saveRDS(rf, here("results", "winning_rf.rds"))
```
