==> analysis/11_testing-tables.Rmd <==
---
title: "Making Table 1"
author: "Ericka B. Smith"
output: pdf_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
library(kableExtra)
library(here)
library(tidyverse)
library(tidyr)
library(magick)
library(readr)
```


```{r}
table_dat <- readRDS(here("results", "table_3_equivalent.rds"))
scene_dat <- read_csv(here("data", "scene_info.csv"))
```

```{r fig.width = 4}
scene_dat[,2:3] %>%
kbl(caption = "Acquisition Dates of Each Landsat 8 Scene", linesep='',booktabs=TRUE,escape=FALSE) %>%
  #kable_material_dark(full_width=F) %>%
  kable_styling(full_width = F, font_size = 14) %>%
   save_kable(file ="../results/tables/scenes.png")
```


```{r fig.width=6}
tibble(lcz = factor(c("7", "7", "9", "9", "15", "15", "16", "16")),#,
       tt = factor(c("train", "test", "train", "test", "train", "test", "train", "test")), 
                 total_n_pixels = rep(0, times=8), 
                 n_polygons = rep(0, times=8)) %>%
  bind_rows(table_dat) %>%
  mutate(lcz = fct_relevel(lcz, c("1", "2","3", "4","5", "6","7", "8","9", "10","11", "12","13", "14","15", "16","17"))) %>%
  arrange(lcz) %>%
  ungroup() %>%
  pivot_wider(names_from=tt, values_from=c(total_n_pixels, n_polygons)) %>%
  unite("Train", c(n_polygons_train, total_n_pixels_train), sep = "        (") %>%
  unite("Test", c(n_polygons_test, total_n_pixels_test), sep = " (") %>%
  mutate(Train = paste(Train, ")", sep=""),
         Test = paste(Test, ")", sep="")) %>%
  relocate(Test, .after=Train) %>%
  mutate(lcz = fct_recode(lcz, "Class 1: Compact high-rise" = "1",
                     "Class 2: Compact mid-rise" = "2",
                     "Class 3: Compact low-rise" = "3",
                     "Class 4: Open high-rise" = "4",
                     "Class 5: Open mid-rise" = "5",
                     "Class 6: Open low-rise" = "6",
                     "Class 7: Lightweight low-rise" = "7",
                     "Class 8: Large low-rise" = "8",
                     "Class 9: Sparsely built" = "9",
                     "Class 10: Heavy Industry" = "10",
                     "Class 11: Dense trees" = "11",
                     "Class 12: Scattered trees" = "12",
                     "Class 13: Bush, scrub" = "13",
                     "Class 14: Low plants" = "14",
                     "Class 15: Bare rock or paved" = "15",
                     "Class 16: Bare soil or sand" = "16",
                     "Class 17: Water" = "17")) %>%
kbl(caption = "\nDelineation of training and test data by polygon and pixel.",linesep='',booktabs=TRUE,escape=FALSE,
        col.names = linebreak(c("Local Climate Zone", "Train", "Test"))) %>%
  kable_styling(full_width = F, font_size = 14) %>%
  add_footnote("Number of polygons is listed first, with number of pixels in parentheses.") %>%
  
  #kable_material_dark(full_width=F) %>%
  save_kable(file = "../results/tables/polygons.png")
```
